fastsync refactor for snapsync
