+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/sync/fastsync/FastSyncActions.java
@@ -198,6 +198,6 @@ public class FastSyncActions {
         ethContext,
         syncState,
         metricsSystem,
-        currentState.getPivotBlockHeader().get());
+        currentState);
   }
 }
