+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/sync/worldstate/TaskQueueIterator.java
@@ -18,11 +18,11 @@ import org.hyperledger.besu.services.tasks.Task;
 
 import java.util.Iterator;
 
-class TaskQueueIterator implements Iterator<Task<NodeDataRequest>> {
+public class TaskQueueIterator<REQUEST> implements Iterator<Task<REQUEST>> {
 
-  private final WorldDownloadState downloadState;
+  private final WorldDownloadState<REQUEST> downloadState;
 
-  public TaskQueueIterator(final WorldDownloadState downloadState) {
+  public TaskQueueIterator(final WorldDownloadState<REQUEST> downloadState) {
     this.downloadState = downloadState;
   }
 
@@ -32,7 +32,7 @@ class TaskQueueIterator implements Iterator<Task<NodeDataRequest>> {
   }
 
   @Override
-  public Task<NodeDataRequest> next() {
+  public Task<REQUEST> next() {
     return downloadState.dequeueRequestBlocking();
   }
 }
