+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/sync/worldstate/WorldDownloadState.java
@@ -18,14 +18,12 @@ import org.hyperledger.besu.ethereum.core.BlockHeader;
 import org.hyperledger.besu.ethereum.eth.manager.EthScheduler;
 import org.hyperledger.besu.ethereum.eth.manager.task.EthTask;
 import org.hyperledger.besu.ethereum.worldstate.WorldStateStorage;
-import org.hyperledger.besu.ethereum.worldstate.WorldStateStorage.Updater;
-import org.hyperledger.besu.services.tasks.CachingTaskCollection;
 import org.hyperledger.besu.services.tasks.Task;
+import org.hyperledger.besu.services.tasks.TaskCollection;
 import org.hyperledger.besu.util.ExceptionUtils;
 
 import java.time.Clock;
 import java.util.Collections;
-import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CancellationException;
 import java.util.concurrent.CompletableFuture;
@@ -36,26 +34,27 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes;
 
-class WorldDownloadState {
+public abstract class WorldDownloadState<REQUEST> {
   private static final Logger LOG = LogManager.getLogger();
 
-  private final boolean downloadWasResumed;
-  private final CachingTaskCollection<NodeDataRequest> pendingRequests;
-  private final int maxRequestsWithoutProgress;
+  private boolean downloadWasResumed;
+  protected final TaskCollection<REQUEST> pendingRequests;
+
+  protected final int maxRequestsWithoutProgress;
   private final Clock clock;
   private final Set<EthTask<?>> outstandingRequests =
       Collections.newSetFromMap(new ConcurrentHashMap<>());
-  private final CompletableFuture<Void> internalFuture;
-  private final CompletableFuture<Void> downloadFuture;
+  protected CompletableFuture<Void> internalFuture;
+  private CompletableFuture<Void> downloadFuture;
   // Volatile so monitoring can access it without having to synchronize.
-  private volatile int requestsSinceLastProgress = 0;
+  protected volatile int requestsSinceLastProgress = 0;
   private final long minMillisBeforeStalling;
   private volatile long timestampOfLastProgress;
-  private Bytes rootNodeData;
-  private WorldStateDownloadProcess worldStateDownloadProcess;
+  protected Bytes rootNodeData;
+  protected WorldStateDownloadProcess worldStateDownloadProcess;
 
   public WorldDownloadState(
-      final CachingTaskCollection<NodeDataRequest> pendingRequests,
+      final TaskCollection<REQUEST> pendingRequests,
       final int maxRequestsWithoutProgress,
       final long minMillisBeforeStalling,
       final Clock clock) {
@@ -65,6 +64,23 @@ class WorldDownloadState {
     this.pendingRequests = pendingRequests;
     this.maxRequestsWithoutProgress = maxRequestsWithoutProgress;
     this.clock = clock;
+    this.internalFuture = new CompletableFuture<>();
+    this.downloadFuture = new CompletableFuture<>();
+    this.internalFuture.whenComplete(this::cleanup);
+    this.downloadFuture.exceptionally(
+        error -> {
+          // Propagate cancellation back to our internal future.
+          if (error instanceof CancellationException) {
+            this.internalFuture.cancel(true);
+          }
+          return null;
+        });
+  }
+
+  public void reset() {
+    this.timestampOfLastProgress = clock.millis();
+    this.requestsSinceLastProgress = 0;
+    this.downloadWasResumed = true;
     this.internalFuture = new CompletableFuture<>();
     this.downloadFuture = new CompletableFuture<>();
     this.internalFuture.whenComplete(this::cleanup);
@@ -122,23 +138,23 @@ class WorldDownloadState {
     return downloadFuture;
   }
 
-  public synchronized void enqueueRequest(final NodeDataRequest request) {
+  public synchronized void enqueueRequest(final REQUEST request) {
     if (!internalFuture.isDone()) {
       pendingRequests.add(request);
       notifyAll();
     }
   }
 
-  public synchronized void enqueueRequests(final Stream<NodeDataRequest> requests) {
+  public synchronized void enqueueRequests(final Stream<REQUEST> requests) {
     if (!internalFuture.isDone()) {
       requests.forEach(pendingRequests::add);
       notifyAll();
     }
   }
 
-  public synchronized Task<NodeDataRequest> dequeueRequestBlocking() {
+  public synchronized Task<REQUEST> dequeueRequestBlocking() {
     while (!internalFuture.isDone()) {
-      final Task<NodeDataRequest> task = pendingRequests.remove();
+      Task<REQUEST> task = pendingRequests.remove();
       if (task != null) {
         return task;
       }
@@ -156,7 +172,8 @@ class WorldDownloadState {
     this.rootNodeData = rootNodeData;
   }
 
-  public synchronized void requestComplete(final boolean madeProgress) {
+  public synchronized void requestComplete(
+      final boolean madeProgress, final long minMillisBeforeStalling) {
     if (madeProgress) {
       requestsSinceLastProgress = 0;
       timestampOfLastProgress = clock.millis();
@@ -169,49 +186,26 @@ class WorldDownloadState {
     }
   }
 
+  public synchronized void requestComplete(final boolean madeProgress) {
+    requestComplete(madeProgress, minMillisBeforeStalling);
+  }
+
   public int getRequestsSinceLastProgress() {
     return requestsSinceLastProgress;
   }
 
-  private synchronized void markAsStalled(final int maxNodeRequestRetries) {
+  protected synchronized void markAsStalled(final int maxNodeRequestRetries) {
     final String message =
         "Download stalled due to too many failures to retrieve node data (>"
             + maxNodeRequestRetries
             + " requests without making progress)";
+
     final WorldStateDownloaderException e = new StalledDownloadException(message);
     internalFuture.completeExceptionally(e);
   }
 
-  public synchronized boolean checkCompletion(
-      final WorldStateStorage worldStateStorage, final BlockHeader header) {
-    if (!internalFuture.isDone() && pendingRequests.allTasksCompleted()) {
-      if (rootNodeData == null) {
-        enqueueRequest(
-            NodeDataRequest.createAccountDataRequest(
-                header.getStateRoot(), Optional.of(Bytes.EMPTY)));
-        return false;
-      }
-      final Updater updater = worldStateStorage.updater();
-      updater.saveWorldState(header.getHash(), header.getStateRoot(), rootNodeData);
-      updater.commit();
-      internalFuture.complete(null);
-      // THere are no more inputs to process so make sure we wake up any threads waiting to dequeue
-      // so they can give up waiting.
-      notifyAll();
-      LOG.info("Finished downloading world state from peers");
-      return true;
-    } else {
-      return false;
-    }
-  }
-
   public synchronized boolean isDownloading() {
     return !internalFuture.isDone();
-  }
-
-  public synchronized void setWorldStateDownloadProcess(
-      final WorldStateDownloadProcess worldStateDownloadProcess) {
-    this.worldStateDownloadProcess = worldStateDownloadProcess;
   }
 
   public synchronized void notifyTaskAvailable() {
@@ -236,4 +230,12 @@ class WorldDownloadState {
         });
     return downloadFuture;
   }
+
+  public void setWorldStateDownloadProcess(
+      final WorldStateDownloadProcess worldStateDownloadProcess) {
+    this.worldStateDownloadProcess = worldStateDownloadProcess;
+  }
+
+  public abstract boolean checkCompletion(
+      final WorldStateStorage worldStateStorage, final BlockHeader header);
 }
